+++ crates/odnelazm-ingest/Cargo.toml
@@ -28,11 +28,8 @@ tokio = { version = "1.49.0", features = ["macros", "rt-multi-thread"] }
 uuid = { version = "1", features = ["v4"] }
 env_logger = "0.11"
 reqwest = { version = "0.13", default-features = false, features = ["json", "rustls"] }
+clap = { version = "4", features = ["derive", "env"] }
 
 [[bin]]
-name = "ingest"
-path = "src/bin/ingest.rs"
-
-[[bin]]
-name = "enrich"
-path = "src/bin/enrich.rs"
+name = "odnelazm-pipeline"
+path = "src/bin/pipeline.rs"
